requirements.txt
streamlit>=1.28.0
torch>=2.1.0
torchvision>=0.16.0
opencv-python==4.10.0.84
Pillow==10.4.0
python-dotenv==1.0.1
paddlepaddle>=2.5.0
paddleocr>=2.7.0
numpy>=1.24.0
huggingface_hub>=0.24.0

